split the component and moved out the main imports
