+++ common/NumberFormatTime.ts
@@ -1,0 +1,2 @@
+const NumberFormatTime = {}
+export default NumberFormatTime
