+++ common/helper.ts
@@ -1,0 +1,14 @@
+export const createLink = (
+  userEmail: string,
+  messageId: string | null | undefined,
+  isDone: boolean | null | undefined,
+  isDeleted: boolean | null | undefined,
+): string => {
+  if (isDone) {
+    return `/messages/done/${messageId}`;
+  } else if (isDeleted) {
+    return `/messages/deleted/${messageId}`;
+  } else {
+    return `/inbox/${userEmail}/${messageId}`;
+  }
+};
