+++ common/periodTypes.ts
@@ -1,0 +1,7 @@
+export const periodTypes = ['Minute', 'Hour', 'Day', 'Week'];
+export const periodRate = {
+  Minute: 1000 * 60,
+  Hour: 1000 * 60 * 60,
+  Day: 1000 * 60 * 60 * 24,
+  Week: 1000 * 60 * 60 * 24 * 7,
+};
