+++ src/App.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const App = () => {
+  return <div className="has-text-centered"></div>;
+};
+
+export default App;
