+++ src/ChipsInput.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ChipsInput = () => {
+  return <div></div>;
+};
+
+export default ChipsInput;
