+++ src/EventDeleteModal.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const EventDeleteModal = () => {
+  return <div></div>;
+};
+
+export default EventDeleteModal;
